// src/app/services/models/cmessage.ts
export interface CMessageDto {
    messageId: string;
    message: string;
    cMessage: string;
    createdOn: string;
    deleteAfter: string;
    to: string;
    subject: string;
    from: string;
    userId: string;
    userName: string;
    userLoginName: string;
    toUserId: string;
    fromUserId: string;
    emailId: string;
    groupId: string;
    sCode: string;
    rid: string;
  }